+++ vibes/interface_assertions.go
@@ -2,13 +2,13 @@ package vibes
 
 var (
 	_ CapabilityAdapter = (*contextCapability)(nil)
-	_ CapabilityAdapter = (*dbCapability)(nil)
+	_ CapabilityAdapter = (*dbCapabilityAdapter)(nil)
 	_ CapabilityAdapter = (*eventsCapability)(nil)
 	_ CapabilityAdapter = (*jobQueueCapability)(nil)
 )
 
 var (
-	_ CapabilityContractProvider = (*dbCapability)(nil)
+	_ CapabilityContractProvider = (*dbCapabilityAdapter)(nil)
 	_ CapabilityContractProvider = (*eventsCapability)(nil)
 	_ CapabilityContractProvider = (*jobQueueCapability)(nil)
 )